refactor: complete refactor of the polycast-backend and changes to config.js
