+++ polycast-backend/package.json
@@ -1,9 +1,11 @@
 {
   "name": "polycast-backend",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "src/server.js",
   "scripts": {
-    "test": "jest"
+    "start": "node src/server.js",
+    "dev": "nodemon src/server.js",
+    "test": "jest --roots src"
   },
   "keywords": [],
   "author": "",
@@ -12,7 +14,6 @@
   "dependencies": {
     "@ffmpeg-installer/ffmpeg": "^1.1.0",
     "@google/generative-ai": "^0.18.0",
-    "ioredis": "^5.3.2",
     "axios": "^1.8.4",
     "cors": "^2.8.5",
     "dotenv": "^16.4.7",
