+++ polycast-backend/src/services/imageService.js
